3/11/20
Finished storage camera and angles and boxing in. Finished upstairs camera angles and boxing in. Incorporated raycasting for object interaction. Prompts for going up and down the stairs work. Doors can be unlocked from other side (but not saved so that they stay that way yet...). Doors that are locked will say so and won't open.